move containerregistryURL to utils
