+++ nf_core/pipelines/download/singularity.py
@@ -18,11 +18,11 @@ import rich.progress
 import nf_core.utils
 from nf_core.pipelines.download.container_fetcher import ContainerFetcher, ContainerProgress
 from nf_core.pipelines.download.utils import (
-    ContainerRegistryUrls,
     DownloadError,
     intermediate_file,
     intermediate_file_no_creation,
 )
+from nf_core.utils import ContainerRegistryUrls
 
 log = logging.getLogger(__name__)
 stderr = rich.console.Console(
